requirements.txt
python-can~=4.5.0
construct~=2.10.70
